chore: move last rpc files to their place
Also update check-deps to be more user-friendly.
