+++ src/dispatchers/browserContextDispatcher.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-import { BrowserContext } from '../../server/browserContext';
+import { BrowserContext } from '../server/browserContext';
 import { Dispatcher, DispatcherScope, lookupDispatcher } from './dispatcher';
 import { PageDispatcher, BindingCallDispatcher, WorkerDispatcher } from './pageDispatcher';
-import * as channels from '../../protocol/channels';
+import * as channels from '../protocol/channels';
 import { RouteDispatcher, RequestDispatcher } from './networkDispatchers';
-import { CRBrowserContext } from '../../server/chromium/crBrowser';
+import { CRBrowserContext } from '../server/chromium/crBrowser';
 import { CDPSessionDispatcher } from './cdpSessionDispatcher';
 
 export class BrowserContextDispatcher extends Dispatcher<BrowserContext, channels.BrowserContextInitializer> implements channels.BrowserContextChannel {
