+++ src/dispatchers/browserTypeDispatcher.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import { BrowserTypeBase, BrowserType } from '../../server/browserType';
+import { BrowserTypeBase, BrowserType } from '../server/browserType';
 import { BrowserDispatcher } from './browserDispatcher';
-import * as channels from '../../protocol/channels';
+import * as channels from '../protocol/channels';
 import { Dispatcher, DispatcherScope } from './dispatcher';
 import { BrowserContextDispatcher } from './browserContextDispatcher';
