+++ src/dispatchers/dialogDispatcher.ts
@@ -14,18 +14,24 @@
  * limitations under the License.
  */
 
-import { ConsoleMessage } from '../../server/console';
-import * as channels from '../../protocol/channels';
+import { Dialog } from '../server/dialog';
+import * as channels from '../protocol/channels';
 import { Dispatcher, DispatcherScope } from './dispatcher';
-import { createHandle } from './elementHandlerDispatcher';
 
-export class ConsoleMessageDispatcher extends Dispatcher<ConsoleMessage, channels.ConsoleMessageInitializer> implements channels.ConsoleMessageChannel {
-  constructor(scope: DispatcherScope, message: ConsoleMessage) {
-    super(scope, message, 'ConsoleMessage', {
-      type: message.type(),
-      text: message.text(),
-      args: message.args().map(a => createHandle(scope, a)),
-      location: message.location(),
+export class DialogDispatcher extends Dispatcher<Dialog, channels.DialogInitializer> implements channels.DialogChannel {
+  constructor(scope: DispatcherScope, dialog: Dialog) {
+    super(scope, dialog, 'Dialog', {
+      type: dialog.type(),
+      message: dialog.message(),
+      defaultValue: dialog.defaultValue(),
     });
+  }
+
+  async accept(params: { promptText?: string }): Promise<void> {
+    await this._object.accept(params.promptText);
+  }
+
+  async dismiss(): Promise<void> {
+    await this._object.dismiss();
   }
 }
