+++ src/dispatchers/dispatcher.ts
@@ -15,11 +15,11 @@
  */
 
 import { EventEmitter } from 'events';
-import { helper } from '../../server/helper';
-import * as channels from '../../protocol/channels';
-import { serializeError } from '../../protocol/serializers';
-import { createScheme, Validator, ValidationError } from '../../protocol/validator';
-import { assert, debugAssert } from '../../utils/utils';
+import { helper } from '../server/helper';
+import * as channels from '../protocol/channels';
+import { serializeError } from '../protocol/serializers';
+import { createScheme, Validator, ValidationError } from '../protocol/validator';
+import { assert, debugAssert } from '../utils/utils';
 
 export const dispatcherSymbol = Symbol('dispatcher');
