+++ src/dispatchers/downloadDispatcher.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { Download } from '../../server/download';
-import * as channels from '../../protocol/channels';
+import { Download } from '../server/download';
+import * as channels from '../protocol/channels';
 import { Dispatcher, DispatcherScope } from './dispatcher';
 import { StreamDispatcher } from './streamDispatcher';
