+++ src/dispatchers/electronDispatcher.ts
@@ -15,8 +15,8 @@
  */
 
 import { Dispatcher, DispatcherScope, lookupDispatcher } from './dispatcher';
-import { Electron, ElectronApplication, ElectronPage } from '../../server/electron/electron';
-import * as channels from '../../protocol/channels';
+import { Electron, ElectronApplication, ElectronPage } from '../server/electron/electron';
+import * as channels from '../protocol/channels';
 import { BrowserContextDispatcher } from './browserContextDispatcher';
 import { PageDispatcher } from './pageDispatcher';
 import { parseArgument, serializeResult } from './jsHandleDispatcher';
