+++ src/dispatchers/frameDispatcher.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { Frame, NavigationEvent } from '../../server/frames';
-import * as channels from '../../protocol/channels';
+import { Frame, NavigationEvent } from '../server/frames';
+import * as channels from '../protocol/channels';
 import { Dispatcher, DispatcherScope, lookupNullableDispatcher, existingDispatcher } from './dispatcher';
 import { ElementHandleDispatcher, createHandle } from './elementHandlerDispatcher';
 import { parseArgument, serializeResult } from './jsHandleDispatcher';
