+++ src/dispatchers/pageDispatcher.ts
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-import { BrowserContext } from '../../server/browserContext';
-import { Frame } from '../../server/frames';
-import { Request } from '../../server/network';
-import { Page, Worker } from '../../server/page';
-import * as channels from '../../protocol/channels';
+import { BrowserContext } from '../server/browserContext';
+import { Frame } from '../server/frames';
+import { Request } from '../server/network';
+import { Page, Worker } from '../server/page';
+import * as channels from '../protocol/channels';
 import { Dispatcher, DispatcherScope, lookupDispatcher, lookupNullableDispatcher } from './dispatcher';
-import { parseError, serializeError } from '../../protocol/serializers';
+import { parseError, serializeError } from '../protocol/serializers';
 import { ConsoleMessageDispatcher } from './consoleMessageDispatcher';
 import { DialogDispatcher } from './dialogDispatcher';
 import { DownloadDispatcher } from './downloadDispatcher';
@@ -28,8 +28,8 @@ import { FrameDispatcher } from './frameDispatcher';
 import { RequestDispatcher, ResponseDispatcher, RouteDispatcher } from './networkDispatchers';
 import { serializeResult, parseArgument } from './jsHandleDispatcher';
 import { ElementHandleDispatcher, createHandle } from './elementHandlerDispatcher';
-import { FileChooser } from '../../server/fileChooser';
-import { CRCoverage } from '../../server/chromium/crCoverage';
+import { FileChooser } from '../server/fileChooser';
+import { CRCoverage } from '../server/chromium/crCoverage';
 
 export class PageDispatcher extends Dispatcher<Page, channels.PageInitializer> implements channels.PageChannel {
   private _page: Page;
