+++ src/dispatchers/playwrightDispatcher.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-import { Playwright } from '../../server/playwright';
-import * as channels from '../../protocol/channels';
+import { Playwright } from '../server/playwright';
+import * as channels from '../protocol/channels';
 import { BrowserTypeDispatcher } from './browserTypeDispatcher';
 import { Dispatcher, DispatcherScope } from './dispatcher';
 import { SelectorsDispatcher } from './selectorsDispatcher';
-import { Electron } from '../../server/electron/electron';
+import { Electron } from '../server/electron/electron';
 import { ElectronDispatcher } from './electronDispatcher';
-import { DeviceDescriptors } from '../../server/deviceDescriptors';
+import { DeviceDescriptors } from '../server/deviceDescriptors';
 
 export class PlaywrightDispatcher extends Dispatcher<Playwright, channels.PlaywrightInitializer> implements channels.PlaywrightChannel {
   constructor(scope: DispatcherScope, playwright: Playwright) {
