+++ src/server.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-import { Transport } from '../protocol/transport';
-import { DispatcherConnection } from './server/dispatcher';
-import { Playwright } from '../server/playwright';
-import { PlaywrightDispatcher } from './server/playwrightDispatcher';
-import { Electron } from '../server/electron/electron';
-import { gracefullyCloseAll } from '../server/processLauncher';
+import { Transport } from './protocol/transport';
+import { DispatcherConnection } from './dispatchers/dispatcher';
+import { Playwright } from './server/playwright';
+import { PlaywrightDispatcher } from './dispatchers/playwrightDispatcher';
+import { Electron } from './server/electron/electron';
+import { gracefullyCloseAll } from './server/processLauncher';
 
 const dispatcherConnection = new DispatcherConnection();
 const transport = new Transport(process.stdout, process.stdin);
